Allow adding bodies through admin area
Admin body edit goes back to admin root with back button.

BodyView used in new image component.

Creating a body is a little intense in that it saves for hte file, then
to create the body, then to save the body data, then to get all bodies,
then to get all files. This is just so that the state has the current
information. Could save a trip to the server by just putting the new
data in the store when it has it.